AgenticAI/PhiData-project-1/requirements.txt
phidata
python-dotenv
yfinance
packaging
googlesearch-python
pycountry
fastapi
uvicorn
groq
openai
